fstests: add support for testing NFS
This adds support for running fstests over NFS either to a preexisting
server or to one spun up by kdevops. The default is to select the
kdevops-deployed NFS server host, but it should also be possible to run
this vs. an arbitrary NFS server by setting the right Kconfig options.

oscheck.sh is modified to mount up the given server and export and then
create $(hostname)/{test,scratch} directories that fstests will use as
the TEST_DEV and SCRATCH_DEV mounts.

By default, it only tests the default mount options, but I also added a
section for enable testing NFSv3. It should be possible to add test sections
for other option combinations in the future.

+++ workflows/fstests/scripts/gendisks.sh
@@ -176,8 +176,12 @@ else
 	if [[ "$TEST_RTDEV" != "" ]]; then
 		RT_MKFS_OPTS="-r rtdev=$TEST_RTDEV"
 	fi
-	echo "$(basename $0) mkfs for TEST_DEV:"
-	CMD="mkfs.$FSTYP $MKFS_OPTIONS $RT_MKFS_OPTS $LOG_MKFS_OPTS $TEST_DEV"
-	echo "$CMD"
-	$CMD
+	if [ $FSTYP = "nfs" ]; then
+		echo "$(basename $0) skipping mkfs"
+	else
+		echo "$(basename $0) mkfs for TEST_DEV:"
+		CMD="mkfs.$FSTYP $MKFS_OPTIONS $RT_MKFS_OPTS $LOG_MKFS_OPTS $TEST_DEV"
+		echo "$CMD"
+		$CMD
+	fi
 fi
